backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic==2.7.1
python-multipart==0.0.9